% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCMoE-package.R
\docType{package}
\name{VCMoE-package}
\alias{VCMoE}
\alias{VCMoE-package}
\title{VCMoE: Varying-Coefficient Mixture-of-Experts Models}
\description{
Fit and evaluate single-response Gaussian, Binomial, and Negative-Binomial
varying-coefficient mixture-of-experts models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qc-zhao.github.io/VCMoE/}
  \item \url{https://github.com/qc-zhao/VCMoE}
  \item Report bugs at \url{https://github.com/qc-zhao/VCMoE/issues}
}

}
\author{
\strong{Maintainer}: Qicheng Zhao \email{790471666@qq.com}

}
\keyword{internal}
